third_party/blink/web_tests/external/wpt/fledge/tentative/round-a-value.https.sub.window-expected.txt

This is a testharness.js-based test.
PASS Check adCost is stochastically rounded with 8 bit mantissa and exponent.
FAIL Check bid is stochastically rounded with 8 bit mantissa and exponent. assert_array_equals: expected property 0 to be "https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_1" but got "https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=seller_report_error" (expected array ["https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_1"] got ["https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=seller_report_error"])
FAIL Check desirability is stochastically rounded with 8 bit mantissa and exponent. assert_array_equals: expected property 0 to be "https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_1" but got "https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=seller_report_error" (expected array ["https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_1"] got ["https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=seller_report_error"])
FAIL Check highestScoringOtherBid is stochastically rounded with 8 bit mantissa and exponent. assert_array_equals: expected property 0 to be "https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_1" but got "https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=seller_report_error" (expected array ["https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=bidder_report_1"] got ["https://web-platform.test:8444/fledge/tentative/resources/request-tracker.py?uuid=<uuid>&dispatch=track_get&id=seller_report_error"])
PASS Value is ignored as a non-valid floating-point number.
PASS Value is rounded to 0 if value is greater than 0 and its exponent is less than -128.
PASS Value is rounded to -0 if value is greater than 0 and its exponent is less than -128.
PASS Value is rounded to Infinity if value is greater than 0 and its exponent is greater than 127.
PASS Value is rounded to -Infinity if value is less than 0 and its exponent is greater than 127.
Harness: the test ran to completion.
